Add "error" event to RosNode and Subscriber, handle tcp stream corruption
This PR refactors error handling to bubble up all failures or unexpected to conditions to either a Subscriber instance or RosNode instance, which will both emit an "error" event. Errors are handled internally, so the error event is just for logging or surfacing information to users.

TCP stream corruption is now detected using the same check in ros_comm. If a length prefix for a rosmsg of >1GB is decoded, assume the stream is corrupted and disconnect.