try to use localAddress for requestTopics response (#26)
ROS subscriber clients use the address and port response arguments to
determine where to establish a connection to receive messages. ROS
clients may be on the local machine or on remote hosts. To better
support subscribers on local or remote hosts, we attempt to use the
socket localAddress of the xmlrpc request if it present. Since the
xmlrpc server and tcp socket publishers are launched within the same
node, echoing the localAddress back as the tpcros destination address
increases the chance the client will be able to establish a connection
to the publishing node since it has already made a successful xmlrpc
request to the same address.